start putting handy test code in LarEditCheckSpec
It's nice to have a way to say that a certain LAR should pass/fail.
It may also be nice to have ways to create those LARs more generally,
but for now, that code is still down in RateSpreadEditCheckSpec.

Specs for objects of type EditCheck[Iterable[LoanApplicationRegister]]
now extend MultipleLarEditCheckSpec, which currently has just one line.
Also, that line is duplicated with LarEditCheckSpec, so perhaps they
could have a common parent instead. I didn't do it now because it feels
like a big hierarchy for not much code, but it may make sense later.