WIP: Authorize requests within each route
Previously, we authorized requests before matching them to a route -- at
the top level of the /institutions and /lar APIs. This resulted in
issues where the rejection response would happen too often (including
for routes that don't exist and should have returned 404).

With this commit, we first match a request to a route (or return 404 if
it doesn't match an endpoint), then authorize. We now use a
RejectionHandler to specifically handle authorization failure with a
403.

TODO: find a way to test the new rejection handling strategy