Remove any trailing newlines before parsing
Previously, since lines are persisted with a newline character at the
end, the TS parser could not remove extra fields (extra | characters)
from the end of the TS CSV. This would cause valid Transmittal Sheets to
fail parsing.
Now, we remove any extra newline characters, and any extra | characters
are removed when the line is split.