Prevent 500 error in field value lookup
I found an inconsistency in the edit-metadata file. "Date
Application Received" is used as another name for "Loan Application
Date". When unhandled, this causes a "key not found" error in Scala and
a 500 in the API.

Now, the field mapping responds correctly for both field names, and it
has a failsafe for cases when the field name doesn't match any in the
mapping.