Start HmdaAPORPersistence actor even if scheduler errors
This error was causing a stacktrace in the logs, and was keeping the
HmdaAPORPersistence actor from starting:
"Unable to store Job : 'DEFAULT.AporCalculator_Job', because one already
exists with this identification."

Instead of stopping when we see that error, continue, since the job
seems to be scheduled already. We may need to add code to remove the old
schedule and make a new one; trying this for now.